modularize view header (wip)
